LPD-8518 refactor AMImageConfigurationEntryParser change the parser class to a util one, and apply to usages and test cases
