+++ modules/apps/adaptive-media/adaptive-media-image-impl/src/test/java/com/liferay/adaptive/media/image/internal/configuration/AMImageConfigurationEntryParserUtilTest.java
@@ -6,6 +6,7 @@
 package com.liferay.adaptive.media.image.internal.configuration;
 
 import com.liferay.adaptive.media.image.configuration.AMImageConfigurationEntry;
+import com.liferay.adaptive.media.image.internal.util.AMImageConfigurationEntryParserUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 
@@ -13,7 +14,6 @@ import java.util.Collections;
 import java.util.Map;
 
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
@@ -21,23 +21,17 @@ import org.junit.Test;
 /**
  * @author Adolfo Pérez
  */
-public class AMImageConfigurationEntryParserTest {
+public class AMImageConfigurationEntryParserUtilTest {
 
 	@ClassRule
 	@Rule
 	public static final LiferayUnitTestRule liferayUnitTestRule =
 		LiferayUnitTestRule.INSTANCE;
 
-	@Before
-	public void setUp() {
-		_amImageConfigurationEntryParser =
-			new AMImageConfigurationEntryParser();
-	}
-
 	@Test
 	public void testDisabledValidString() {
 		AMImageConfigurationEntry amImageConfigurationEntry =
-			_amImageConfigurationEntryParser.parse(
+			AMImageConfigurationEntryParserUtil.parse(
 				"test:desc:12345:max-height=100;max-width=200:enabled=false");
 
 		Assert.assertEquals("test", amImageConfigurationEntry.getName());
@@ -55,36 +49,36 @@ public class AMImageConfigurationEntryParserTest {
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testEmptyAttributes() {
-		_amImageConfigurationEntryParser.parse("test:desc:12345:");
+		AMImageConfigurationEntryParserUtil.parse("test:desc:12345:");
 	}
 
 	@Test
 	public void testEmptyDescription() {
-		_amImageConfigurationEntryParser.parse(
+		AMImageConfigurationEntryParserUtil.parse(
 			"test::12345:max-height=100;max-width=200");
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testEmptyName() {
-		_amImageConfigurationEntryParser.parse(
+		AMImageConfigurationEntryParserUtil.parse(
 			":desc:12345:max-height=100;max-width=200");
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testEmptyString() {
-		_amImageConfigurationEntryParser.parse("");
+		AMImageConfigurationEntryParserUtil.parse("");
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testEmptyUUID() {
-		_amImageConfigurationEntryParser.parse(
+		AMImageConfigurationEntryParserUtil.parse(
 			"test:desc::max-height=100;max-width=200");
 	}
 
 	@Test
 	public void testEncodedDescription() {
 		AMImageConfigurationEntry amImageConfigurationEntry =
-			_amImageConfigurationEntryParser.parse(
+			AMImageConfigurationEntryParserUtil.parse(
 				"test:desc%3A%3B:12345:max-height=100;max-width=200");
 
 		Assert.assertEquals("test", amImageConfigurationEntry.getName());
@@ -103,7 +97,7 @@ public class AMImageConfigurationEntryParserTest {
 	@Test
 	public void testEncodedName() {
 		AMImageConfigurationEntry amImageConfigurationEntry =
-			_amImageConfigurationEntryParser.parse(
+			AMImageConfigurationEntryParserUtil.parse(
 				"test%3A%3B:desc:12345:max-height=100;max-width=200");
 
 		Assert.assertEquals("test:;", amImageConfigurationEntry.getName());
@@ -130,7 +124,7 @@ public class AMImageConfigurationEntryParserTest {
 
 		Assert.assertEquals(
 			"test:desc:12345:max-height=100:enabled=true",
-			_amImageConfigurationEntryParser.getConfigurationString(
+			AMImageConfigurationEntryParserUtil.getConfigurationString(
 				amImageConfigurationEntry));
 	}
 
@@ -148,7 +142,7 @@ public class AMImageConfigurationEntryParserTest {
 
 		Assert.assertEquals(
 			"test:desc:12345:max-height=100;max-width=200:enabled=true",
-			_amImageConfigurationEntryParser.getConfigurationString(
+			AMImageConfigurationEntryParserUtil.getConfigurationString(
 				amImageConfigurationEntry));
 	}
 
@@ -164,7 +158,7 @@ public class AMImageConfigurationEntryParserTest {
 
 		Assert.assertEquals(
 			"test:desc:12345:max-width=200:enabled=true",
-			_amImageConfigurationEntryParser.getConfigurationString(
+			AMImageConfigurationEntryParserUtil.getConfigurationString(
 				amImageConfigurationEntry));
 	}
 
@@ -176,7 +170,7 @@ public class AMImageConfigurationEntryParserTest {
 
 		Assert.assertEquals(
 			"test:desc:12345::enabled=true",
-			_amImageConfigurationEntryParser.getConfigurationString(
+			AMImageConfigurationEntryParserUtil.getConfigurationString(
 				amImageConfigurationEntry));
 	}
 
@@ -192,7 +186,7 @@ public class AMImageConfigurationEntryParserTest {
 
 		Assert.assertEquals(
 			"test:desc:12345:max-height=100:enabled=false",
-			_amImageConfigurationEntryParser.getConfigurationString(
+			AMImageConfigurationEntryParserUtil.getConfigurationString(
 				amImageConfigurationEntry));
 	}
 
@@ -210,7 +204,7 @@ public class AMImageConfigurationEntryParserTest {
 
 		Assert.assertEquals(
 			"test:desc:12345:max-height=100;max-width=200:enabled=false",
-			_amImageConfigurationEntryParser.getConfigurationString(
+			AMImageConfigurationEntryParserUtil.getConfigurationString(
 				amImageConfigurationEntry));
 	}
 
@@ -226,31 +220,31 @@ public class AMImageConfigurationEntryParserTest {
 
 		Assert.assertEquals(
 			"test:desc:12345:max-width=200:enabled=false",
-			_amImageConfigurationEntryParser.getConfigurationString(
+			AMImageConfigurationEntryParserUtil.getConfigurationString(
 				amImageConfigurationEntry));
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testInvalidEnabledAttribute() {
-		_amImageConfigurationEntryParser.parse(
+		AMImageConfigurationEntryParserUtil.parse(
 			"test:desc:12345:max-height=100;max-width=200:disabled=true");
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testMissingAttributesField() {
-		_amImageConfigurationEntryParser.parse("test:desc:12345");
+		AMImageConfigurationEntryParserUtil.parse("test:desc:12345");
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testMissingDescription() {
-		_amImageConfigurationEntryParser.parse(
+		AMImageConfigurationEntryParserUtil.parse(
 			"12345:max-height=100;max-width=200");
 	}
 
 	@Test
 	public void testMissingEnabledAttributeDefaultsEnabled() {
 		AMImageConfigurationEntry amImageConfigurationEntry =
-			_amImageConfigurationEntryParser.parse(
+			AMImageConfigurationEntryParserUtil.parse(
 				"test:desc:12345:max-height=100;max-width=200");
 
 		Assert.assertEquals("test", amImageConfigurationEntry.getName());
@@ -268,25 +262,25 @@ public class AMImageConfigurationEntryParserTest {
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testMissingName() {
-		_amImageConfigurationEntryParser.parse(
+		AMImageConfigurationEntryParserUtil.parse(
 			"12345:desc:max-height=100;max-width=200");
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testMissingUUID() {
-		_amImageConfigurationEntryParser.parse(
+		AMImageConfigurationEntryParserUtil.parse(
 			"test:desc:max-height=100;max-width=200");
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void testNullString() {
-		_amImageConfigurationEntryParser.parse(null);
+		AMImageConfigurationEntryParserUtil.parse(null);
 	}
 
 	@Test
 	public void testValidString() {
 		AMImageConfigurationEntry amImageConfigurationEntry =
-			_amImageConfigurationEntryParser.parse(
+			AMImageConfigurationEntryParserUtil.parse(
 				"test:desc:12345:max-height=100;max-width=200:enabled=true");
 
 		Assert.assertEquals("test", amImageConfigurationEntry.getName());
@@ -301,7 +295,5 @@ public class AMImageConfigurationEntryParserTest {
 		Assert.assertEquals("200", properties.get("max-width"));
 		Assert.assertEquals(properties.toString(), 2, properties.size());
 	}
-
-	private AMImageConfigurationEntryParser _amImageConfigurationEntryParser;
 
 }
